+++ tests/common.js
@@ -1,6 +1,7 @@
 var path = require("path");
 require.paths.unshift(path.dirname(__dirname)+'/lib');
 global.Gently = require("gently");
+global.GENTLY_HIJACK = new Gently();
 global.assert = require("assert");
 global.ntest = require('ntest');
 global.it = ntest.it;