+++ tests/index.js
@@ -3,6 +3,7 @@ require("./lastfm-recenttracks-parser-test.js");
 require("./lastfm-recenttracks-stream-test.js");
 require("./lastfm-session-test.js");
 require("./lastfm-info-test.js");
+require("./lastfm-read-test.js");
 require("./lastfm-info-track-test.js");
 require("./lastfm-update-test.js");
 require("./lastfm-request-test.js");