+++ tests/lastfm-info-test.js
@@ -11,54 +11,55 @@ describe("a new info instance")
     gently = new Gently();
   });
 
-  it("accepts listeners in options", function() {
+  it("accepts listeners in options (deprecated)", function() {
+      var handlers = { error: function() {}, success: function() {} };
+      gently.expect(handlers, "error");
+      gently.expect(handlers, "success");
+      var info = new LastFmInfo(lastfm, "", handlers);
+      info.emit("success");
+    });
+
+  it("accepts listeners in handler options", function() {
     var handlers = { error: function() {}, success: function() {} };
+    var options = { handlers: handlers };
     gently.expect(handlers, "error");
     gently.expect(handlers, "success");
-    var info = new LastFmInfo(lastfm, "", handlers);
+    var info = new LastFmInfo(lastfm, "", options);
     info.emit("success");
   });
   
   it("emits error if type not specified", function() {
-    var handler = { error: function() {}};
-    gently.expect(handler, "error", function(error) {
+    var handlers = { error: function() {}};
+    gently.expect(handlers, "error", function(error) {
       assert.equal("Item type not specified", error.message);
     });
-    var info = new LastFmInfo(lastfm, "", { error: handler.error });
+    var info = new LastFmInfo(lastfm, "", { handlers: handlers });
   });
   
   it("allows requests for user info", function() {
-    gently.expect(lastfm, "read", function() {
+    gently.expect(lastfm, "request", function() {
       return new fakes.LastFmRequest();
     });
     var info = new LastFmInfo(lastfm, "user");
   });
 
   it("allows requests for track info", function() {
-    gently.expect(lastfm, "read", function() {
+    gently.expect(lastfm, "request", function() {
       return new fakes.LastFmRequest();
     });
     var info = new LastFmInfo(lastfm, "track");
   });
 
   it("allows all [itemtype].getinfo calls", function() {
-    gently.expect(lastfm, "read", function(params) {
-      assert.equal("event.getinfo", params.method);
+    gently.expect(lastfm, "request", function(method, params) {
+      assert.equal("event.getinfo", method);
       return new fakes.LastFmRequest();
     });
     new LastFmInfo(lastfm, "event");
   });
   
-  it("calls unsigned methods", function() {
-    gently.expect(lastfm, "read", function(params, signed) {
-      assert.equal(false, signed);
-      return new fakes.LastFmRequest();
-    });
-    var info = new LastFmInfo(lastfm, "user");
-  });
-
   it("passes through parameters", function() {
-    gently.expect(lastfm, "read", function(params) {
+    gently.expect(lastfm, "request", function(method, params) {
       assert.equal("username", params.user);
       assert.equal("anything", params.arbitrary);
       return new fakes.LastFmRequest();
@@ -67,12 +68,12 @@ describe("a new info instance")
   });
 
   it("doesnt pass through callback parameters", function() {
-    gently.expect(lastfm, "read", function(params) {
+    gently.expect(lastfm, "request", function(method, params) {
       assert.ok(!params.error);
       assert.ok(!params.success);
       return new fakes.LastFmRequest();
     });
-    new LastFmInfo(lastfm, "user", { error: function() {}, success: function() {} });
+    new LastFmInfo(lastfm, "user", { handlers:  { error: function() {}, success: function() {} } });
   });
 })();
 
@@ -86,57 +87,57 @@ describe("when receiving data")
   });
 
   it("emits error if response contains error", function() {
-    gently.expect(lastfm, "read", function(params, signed) {
+    gently.expect(lastfm, "request", function() {
       return request;
     });
-    new LastFmInfo(lastfm, "track", {
+    new LastFmInfo(lastfm, "track", { handlers: {
       error: gently.expect(function errorHandler(error) {
         assert.equal("You must supply either a track & artist name or a track mbid.", error.message);
       })
-    });
+    }});
     request.emit("success", FakeData.NotEnoughTrackInfo);
   });
 
   it("emits error when receiving unexpected data", function() {
-    gently.expect(lastfm, "read", function(params, signed) {
+    gently.expect(lastfm, "request", function() {
       return request;
     });
-    new LastFmInfo(lastfm, "track", {
+    new LastFmInfo(lastfm, "track", { handlers: {
       error: gently.expect(function errorHandler(error) {
         assert.equal("Unexpected error", error.message);
       })
-    });
+    }});
     request.emit("success", FakeData.SuccessfulAuthorisation);
   });
 
   it("emits error if receiving junk", function() {
-      gently.expect(lastfm, "read", function(params, signed) {
+      gently.expect(lastfm, "request", function() {
         return request;
       });
-      new LastFmInfo(lastfm, "track", {
+      new LastFmInfo(lastfm, "track", { handlers: {
         error: gently.expect(function errorHandler(error) {
           assert.ok(error.message.indexOf(FakeData.Garbage) > -1);
           assert.ok("Syntax error");
         })
-      });
+      }});
       request.emit("success", FakeData.Garbage);
   });
 
   it("emits success with received data when matches expected type", function() {
-    gently.expect(lastfm, "read", function(params, signed) {
+    gently.expect(lastfm, "request", function() {
         return request;
     });
-    new LastFmInfo(lastfm, "track", {
+    new LastFmInfo(lastfm, "track", { handlers: {
       success: gently.expect(function success(track) {
         assert.equal("Run To Your Grave", track.name);
         assert.equal("232000", track.duration);
       })
-    });
+    }});
     request.emit("success", FakeData.RunToYourGraveTrackInfo);
   });
 
   it("bubbles up errors", function() {
-    gently.expect(lastfm, "read", function(params, signed) {
+    gently.expect(lastfm, "request", function() {
         return request;
     });
     var info = new LastFmInfo(lastfm, "track");